+++ Mailer/extensions/mailer-module/src/migrations/20230525081510_add_mail_types_foreign_key_mail_type_stats.php
@@ -1,0 +1,14 @@
+<?php
+declare(strict_types=1);
+
+use Phinx\Migration\AbstractMigration;
+
+final class AddMailTypesForeignKeyMailTypeStats extends AbstractMigration
+{
+    public function change(): void
+    {
+        $this->table('mail_type_stats')
+            ->addForeignKey('mail_type_id', 'mail_types', 'id')
+            ->update();
+    }
+}