+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - Added copy banner link to banner detail and banner edit form. remp/remp#1259
 
+### [Mailer]
+
+- Removed deleted mail types from dashboard stats. remp/remp#1269
+
 ## Archive
 
 - [v2.2](./changelogs/CHANGELOG-v2.2.md)